Attemped to use collision detection technique.
The technique is spliting the screen in areas.

This way, when storing a dot, we'd have to check intersection of 10
rects, but when drawing we'd be only checking for the circles on the
area. This shoudl improve performance for a high number of circles.

Still this is not working on the borders, since the circles are being
drawn in an incorrect order.